refactor: 타입 변환 getPathsDistance 메소드에서 하도록 변경

+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -26,7 +26,7 @@ public class PathService {
         Station target = stationRepository.findById(targetId).orElseThrow(() -> new IllegalArgumentException("도착역이 존재하지 않습니다."));
         PathFinder pathFinder = new PathFinder(lineRepository.findAll());
         List<StationResponse> shortestPaths = pathFinder.findPaths(source, target);
-        int pathsDistance = (int) pathFinder.getPathsDistance(source, target);
+        int pathsDistance = pathFinder.getPathsDistance(source, target);
         return new PathResponse(shortestPaths, pathsDistance);
     }
 }
